_v02/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.6
anthropic>=0.40
openpyxl>=3.1
python-docx>=1.1
msal>=1.26
pyyaml>=6.0
python-dotenv>=1.0
httpx>=0.27
botbuilder-core>=4.14
botbuilder-integration-aiohttp>=4.14
aiohttp>=3.9
pydantic>=2.5
